requirements.txt
streamlit
pandas
yfinance
openpyxl
google-genai
python-dotenv